Fix file-sync album matching
Since for some reason I thought it was a good idea to match only _some_ of the filters instead of all of them, doing something like adding the Spotify provider first and then a file-based one, would result in the first album of a band present in both libraries being partially overriden by the file-based sync and causing a bunch of funny issues. This hopefully solves that!