Fix end of track detection for Spotify
Well, this is getting fun seeing it broken every few months! I don't know when this broke, but
seems like the only way of properly detecting this now requires knowing about the previous state
plus some funky debouncing logic. That coupled with the way we're pushing commands onto the media
player event queue and it makes this soup we see here!

Damn you Spotify for making it so hard to rely on this logic. It seems you want to make developer's
life as hard as you're making musician's :)