Improve download/parsing error reporting
For some reason Safari refuses to work on prod even though it works on my computer (a classic!), but the errors were extremely hidden by the choice of error data type I chose at the time. So changing them to actually bring the error along instead of losing it on the way and *hopefully* that'll give me some hint of what's going on.